Added favicons html and include ot base.html looks for icons in /fav/
This is the standard configuration that works well with http://realfavicongenerator.net
favicons.html is currently set to use /favs/ as the directory where it will tell
the browser to look for favicons. In the future this can possibly be abstracted
out to be a variable. base.html has been edited to include favicons.html inside
the <head> tag.